# Improve multi-thread performance by releasing GVL