… ext/RMagick/rmilist.c
… ext/RMagick/rmimage.c