With releasing GVL, it would improve performance when processing images in a multi-thread.

Fix https://github.com/rmagick/rmagick/issues/243

## Sample
```ruby
require 'rmagick'

t1 = Thread.new do
  1000.times do
    img = Magick::Image.read('Flower_Hat.jpg').first
    thumbnail = img.resize(76, 76)
  end
end

t2 = Thread.new do
  1000.times do
    img = Magick::Image.read('Flower_Hat.jpg').first
    thumbnail = img.resize(76, 76)
  end
end

t1.join
t2.join
```


Before | After
-- | --
3.455 sec | 1.766 sec


### Environment
- Linux
  - Manjaro Linux x86_64
  - Kernel: 6.0.2-2-MANJARO
  - AMD Ryzen 7 5800H
  - gcc version 12.2.0
  - Ruby 3.1.2

### Before

```
$ time ruby thread.rb
ruby thread.rb  3.30s user 0.09s system 98% cpu 3.455 total
```

### After

```
$ time ruby thread.rb
ruby thread.rb  3.32s user 0.07s system 192% cpu 1.766 total
```