@mockdeep 2022-10-30
@Watson1978 I don't think I'm qualified to review this. Is it a risky change? Should we release it under a beta?

Is there anyone in the core ruby community who can look it over? @dlemstra is this the sort of thing you can look over?